enterprise-logistics/services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic[email]
bcrypt==3.2.2
python-multipart
pytest
httpx
pytest-cov
